requirements.txt
streamlit==1.36.0
pandas
plotly-express